requirements.txt
asgiref==3.5.1
Django==4.0.4
python-decouple==3.6
sqlparse==0.4.2
line-bot-sdk==2.2.1